move test code outside of lib
